bugfix: handle case if path from enemy to player not exists
