+++ maps/map1.ber
@@ -1,6 +1,6 @@
 11111111111111111
-1E000000C000000C1
-10100101001000001
-10100100101010101
+1E000010C000000C1
+10100111001000001
+10100110101010101
 1P000000000S00001
 11111111111111111
